Refactor popularity score calculation and move related logic to new module
- Removed the RADIUS_ZOOM_MULTIPLIER and calculate_category_multiplier functions from data_fetcher.py.
- Introduced a new module, popularity_algo.py, to encapsulate the radius zoom multiplier and category multiplier logic.
- Updated the new_ggl_to_boxmap method in mapbox_connector.py to accept a radius parameter and calculate popularity scores based on the new module.
- Enhanced the fetch_ggl_nearby function to utilize the updated MapBoxConnector method for improved data handling.
